// src/types/supabase.ts
export type Database = {
// ...
  public: {
// ...
    Enums: {
      drive: 'FWD' | 'RWD' | '4WD' | 'AWD';
      fuel:
        | 'gasoline'
        | 'diesel'
        | 'CNG'
        | 'LPG'
        | 'hybrid'
        | 'electric'
        | 'hydrogen'
        | 'ethanol';
      service_category:
        | 'battery'
        | 'body'
        | 'brakes'
        | 'electrical'
        | 'engine'
        | 'interior'
        | 'other'
        | 'suspension'
        | 'tires';
      transmission: 'manual' | 'automatic' | 'CVT';
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
// ...
};

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  'public'
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema['Enums']
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;